requirements.txt
pandas
mysql-connector-python 
SQLAlchemy 
openpyxl 
pymysql
fastapi 
uvicorn 
sqlalchemy 
python-jose[cryptography] 
passlib[bcrypt]
python-multipart
fastapi
sqlalchemy
pydantic
python-dotenv
bcrypt
aiofiles
httpx
jinja2
